Extract lifted source owner resolver
