+++ docs/design/method-body-inspection.md
@@ -310,12 +310,20 @@ and delegate facts, async-state-machine caching, and the narrow resolver
 adapters. `LibraryBodyMethodReferenceResolver` owns the acquisition-scoped
 structural signature and generic-scope identities, canonical
 `MemberRef`/`MethodSpec` resolution caches, and their shared assembly work
-budgets. The primary resolver adapters and lifted-source-owner reference index
-consume that same resolution authority.
+budgets. The primary resolver adapters and
+`LibraryBodyLiftedSourceOwnerResolver` consume that same resolution authority.
+The lifted-source-owner resolver owns acquisition-scoped local-function/lambda
+owner correlation, memoized owner-body reference evidence, top-level
+entry-point authentication, and classic async state-machine type-name
+resolution. It reuses primary metadata identity and generated-code judgments.
 `OptimizationOpportunities_DuplicateMemberRefsResolveStructuralIdentityOnce`,
 `OptimizationOpportunities_SharedMemberRefDecodesOnceAcrossOwnerBodies`, and
 `LiftedOwnerMemberIdentity_RetainsExactAssemblyReferenceScope` gate cache
 sharing and scope-aware identity.
+`OptimizationOpportunities_LiftedOwnerBody_IsIndexedOnce`,
+`OptimizationOpportunities_ClassicAsyncTypeDefinitionsAreIndexedOnce`, and the
+top-level local-function tests gate the lifted-owner caches and execution
+mapping.
 `CallerUnsafeMode_PointerSignatureIsImplicitWhenModuleNotOptedIn`,
 `OptimizationOpportunities_AsyncStateMachine_IsAmortized`, and
 `Allocations_ClassifiesCrossAndInAssemblyValueTypeNewobj_ByShape` gate
